feat: NextComponent Store 연결 및 파일명 변경
