+++ src/components/invitation/create/InvitationVisitorsList.tsx
@@ -1,26 +1,26 @@
-import { css } from '@emotion/react';
-import NameTag from '@/components/common/NameTag';
 import Add from '@/components/common/Add';
+import NameTag from '@/components/common/NameTag';
+import CREATE_TEXTS from '@/constants/invitation/createTexts';
 import mq from '@/utils/mediaquery';
 import theme from '@/styles/theme';
-import CREATE_TEXTS from '@/constants/invitation/createTexts';
+import { css } from '@emotion/react';
 import {
-  InvitationNameListProps,
+  InvitationVisitorsListProps,
   InvitationCreateTexts,
 } from '@/types/invitation/create';
 
-function InvitationNameList({ nameList }: InvitationNameListProps) {
+function InvitationVisitorsList({ visitorsList }: InvitationVisitorsListProps) {
   const { title }: InvitationCreateTexts = CREATE_TEXTS;
 
   return (
-    <div css={listContainerStyles}>
+    <div css={containerStyles}>
       <div css={titleWrapperStyles}>
         <div css={titleStyles}>{title.invitationList}</div>
-        <div css={lengthStyles}>{nameList.length}/30</div>
+        <div css={lengthStyles}>{visitorsList.length}/30</div>
       </div>
       <div css={listWrapperStyles}>
         <Add isBlue onClick={() => alert('추가 버튼 테스트')} />
-        {nameList.map((name) => (
+        {visitorsList.map((name: string) => (
           <NameTag
             key={name}
             name={name}
@@ -32,7 +32,7 @@ function InvitationNameList({ nameList }: InvitationNameListProps) {
   );
 }
 
-const listContainerStyles = css`
+const containerStyles = css`
   display: flex;
   flex-direction: column;
   gap: 16px;
@@ -74,4 +74,4 @@ const listWrapperStyles = css`
   width: 100%;
 `;
 
-export default InvitationNameList;
+export default InvitationVisitorsList;
